+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- `Nanook::Wallet#restore` to create a wallet, change its seed and create x number of accounts
 - `Nanook::WalletAccount#create` takes an optional argument to signal
   how many accounts to create.
 - New `Nanook::WalletAccount#change_representative` method to change an